feat!: remove outdated defs and lemmas from TM
"extract" Hoare and language related defs and proofs into separate theory TM_Hoare. they may be of use in the future, but for the time being are broken. the current plan is to build new definitions (possible omitting hoare defs if they do not provide much use).

remove the old composition, rejecting TM and natTM, as they either seem not salvageable (too many differences to the new spec) or have already been implemented for the new spec.
remove hoare rules as they are currently unused

adjust the "broken" marker in the ROOT file