fix: TM_comp when M1.q0 is final, minor improvements
if M1.q0 is final, then M.q0 = M2.q0 directly.
previously M would just execute steps from M1, even though it should just proceed to M2.q0.

the behavior is still incorrent when starting execution in arbitrary c where ``state c : Inl ` M1.F``. a general fix is possible, but annoying and complex. if it is not directly necessary, it will be left as is for now.