feat(TM): rework transition function (WIP)
problem: the intended TM encoding should be injective.
this is not actually possible with the current definition; as the types of states and symbols can be infinite, the range of the transition function(s) can be as well.
this is not interesting/useful for reasoning about TMs and causes TMs to have an infinite amount of "variants" that all have the same encoding, but are not "equal" as far as Isabelle's `HOL.eq` (equivalent to `Pure.eq` in HOL) is concerned.

proposed solution: add axioms to valid_TM that fix/define the behavior of the transition function (next_state, next_write, next_move) in the "undefined" range (for non-well-formed/invalid input values).
for two TMs to be equal (`HOL.eq`) it then suffices for their transition functions to have equal results in the intended range.

Note: this approach works due to function extensionability (`thm HOL.ext`)